+++ Lesson_5/Classes_and_Inheritance.py
@@ -1,58 +1,18 @@
-class Block:
+import Block_data
 
-    tooltype = "Pickaxe"
-    num_of_blocks = 0
-    damage = 1
-
-    def __init__(self, name, hardness, resistance, damage, tooltype):
-        self.name = name
-        self.hardness = hardness
-        self.resistance = resistance
-        self.damage = damage
-        self.tooltype = tooltype
-
-        Block.num_of_blocks += 1
-    def __repr__(self) -> str:
-        return "Block('{}', '{}', '{}', '{}', '{}')".format(self.name, self.hardness, self.resistance, self.damage, self.tooltype)
-
-    def __str__(self):
-        return "Block: {}\nHarvest Tool: {}".format(self.name, self.tooltype)
-
-    def getdamage(self): # Getter
-        return self.damage
-
-    def setDamage(self, newDamage): # Setter
-        self.damage = newDamage
-        return "Damage has been set to " + newDamage
-
-
-    @classmethod
-    def light(cls, level):
-        level = 0
-    
-    @classmethod
-    def fromstring(cls, newblock):
-        name, hardness, resistance, damage, tooltype = newblock.split('-')
-        return cls(name, hardness, resistance, damage, tooltype)
-    
     # @staticmethod
     # def time(day):
     #     if day.hour <= 12:
     #         return "Day"
     #     return "Night"
 
-class itemstack(Block):
-    def __init__(self, name, hardness, resistance, damage, tooltype, state):
-        super().__init__(name, hardness, resistance, damage, tooltype)
-        self.state = state
-
-Ruby = Block('Ruby', '2', '3', '1', 'pickaxe')
-Coal = Block('Block of Coal', '5', '6', '1', 'pickaxe')
-Bucket_of_Water = itemstack('Bucket of Water', '0', '0', '1', 'fist', 'liquid')
+Ruby = Block_data.Block('Ruby', '2', '3', '1', 'pickaxe')
+Coal = Block_data.Block('Block of Coal', '5', '6', '1', 'pickaxe')
+Bucket_of_Water = Block_data.itemstack('Bucket of Water', '0', '0', '1', 'fist', 'liquid')
 print(Ruby.__repr__())
 print(Ruby.__str__())
 # createblock = str(input("Name-Hardness-Resistance-Damage-Tooltype\n"))
-# newblock = Block.fromstring(createblock)
+# newblock = Block_data.fromstring(createblock)
 # print(Ruby_Block.name)
 # print(Coal_Block.name)
 
@@ -62,7 +22,7 @@ print(Ruby.__str__())
 
 # the_time = datetime.time(11, 00)
 
-# print(Block.time(the_time))
+# print(Block_data.time(the_time))
 
 # print(Ruby.hardness)
 # print(Coal.resistance)
@@ -72,11 +32,11 @@ print(Ruby.__str__())
 # print(Ruby.setDamage('SKY BLUE'))
 # print(Ruby.getdamage())
 # print(Ruby.tooltype)
-# print(Block.num_of_blocks)
+# print(Block_data.num_of_blocks)
 # print('\n')
 # print(newblock.damage)
-# print(Block.damage(Ruby.damage))
-# print(Block.hardness(Coal.hardness))
+# print(Block_data.damage(Ruby.damage))
+# print(Block_data.hardness(Coal.hardness))
 
 # print(Ruby.damage)
 # print(Coal.hardness)
